No keyboard scrolling in the direction of edges
Now you can't scroll up or down with the Cylinder or Mobius strip
topologies.

TODO: I also noticed that the Mobius strip topology does not actually
seem to work. I'm trying to determine now if it's just the display that
is wrong or if the internal logic of the board is wrong.